feat: lazy load all react-page plugins to reduce bundle size
