+++ packages/core/src/helper/lazyLoad/index.tsx
@@ -1,0 +1,4 @@
+import loadable from '@loadable/component';
+
+// tslint:disable-next-line:no-any
+export default (loader: any) => loadable(loader);
