+++ packages/plugins/content/divider/src/Component/index.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import { DividerProps } from '../types/component';
 
 const Divider: React.SFC<DividerProps> = props => {
-  return <props.Controls {...props}/>;
+  return <props.Renderer {...props} />;
 };
 
 export default Divider;
