--- packages/plugins/content/divider/src/Controls/DividerDefaultControls.tsx
@@ -1,10 +1,0 @@
-import * as React from 'react';
-import { DividerControlsProps } from '../types/controls';
-
-const DividerDefaultControls: React.SFC<DividerControlsProps> = props => {
-    return (
-        <props.Renderer {...props}/>
-    );
-};
-
-export default DividerDefaultControls;
