+++ packages/plugins/content/html5-video/src/Controls/Html5VideoDefaultControls.tsx
@@ -17,24 +17,18 @@ const Html5VideoDefaultControls: React.SFC<Html5VideoControlsProps> = props => {
     commitUrl,
     changeUrlPreview,
     focused,
-    readOnly,
   } = props;
   return (
-    <div className="ory-content-plugin-html5-video">
-      {!readOnly && focused && (
-        <BottomToolbar open={focused} theme={darkTheme}>
-          <TextField
-            placeholder={props.translations.urlPlaceholder}
-            label={props.translations.urlLabel}
-            onChange={e => changeUrlPreview(e.target.value)}
-            onBlur={commitUrl}
-            value={url}
-            style={{ width: '512px' }}
-          />
-        </BottomToolbar>
-      )}
-      <props.Renderer {...props} />
-    </div>
+    <BottomToolbar open={focused} theme={darkTheme}>
+      <TextField
+        placeholder={props.translations.urlPlaceholder}
+        label={props.translations.urlLabel}
+        onChange={e => changeUrlPreview(e.target.value)}
+        onBlur={commitUrl}
+        value={url}
+        style={{ width: '512px' }}
+      />
+    </BottomToolbar>
   );
 };
