+++ packages/plugins/content/html5-video/src/Renderer/Html5VideoHtmlRenderer.tsx
@@ -6,16 +6,18 @@ const Html5VideoHtmlRenderer: React.SFC<Html5VideoRendererProps> = ({
   state: { url = '' } = defaultHtml5VideoState,
 }) => {
   return (
-    <video
-      autoPlay={true}
-      controls={true}
-      loop={true}
-      muted={true}
-      width="100%"
-      key={url}
-    >
-      <source src={url} type={`video/${url.split('.').pop()}`} />
-    </video>
+    <div className="ory-content-plugin-html5-video">
+      <video
+        autoPlay={true}
+        controls={true}
+        loop={true}
+        muted={true}
+        width="100%"
+        key={url}
+      >
+        <source src={url} type={`video/${url.split('.').pop()}`} />
+      </video>
+    </div>
   );
 };
