+++ packages/plugins/content/html5-video/src/createPlugin.tsx
@@ -41,16 +41,13 @@ const rejectPromise: (e: Event, props: Props) => Promise<void> = (
 
 const createPlugin: (
   settings: Html5VideoSettings
-) => ContentPluginConfig<Html5VideoState> = (
-  settings
-) => {
+) => ContentPluginConfig<Html5VideoState> = settings => {
   const mergedSettings = { ...defaultSettings, ...settings };
   const WrappedComponent: React.SFC<Html5VideoProps> = props => (
     <Html5Video {...props} {...mergedSettings} />
   );
   return {
     Component: WrappedComponent,
-    StaticComponent: settings.Renderer,
     name: 'ory/sites/plugin/content/html5-video',
     version: '0.0.1',
     text: mergedSettings.translations.pluginName,
