+++ packages/plugins/content/image/src/Component/index.tsx
@@ -38,9 +38,7 @@ class Form extends React.Component<ImageProps, StateType> {
     this.state = {};
   }
 
-  handleChange = (
-    e: React.ChangeEvent<HTMLInputElement>
-  ) => {
+  handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
     const target = e.target;
     if (target instanceof HTMLInputElement) {
       let change: Partial<ImageState> = {};
@@ -72,15 +70,20 @@ class Form extends React.Component<ImageProps, StateType> {
   }
 
   render() {
-    const { Controls } = this.props;
+    const { Controls, Renderer, readOnly } = this.props;
     return (
-      <Controls
-        {...this.props}
-        imagePreview={this.state.imagePreview}
-        handleImageLoaded={this.handleImageLoaded}
-        handleImageUploaded={this.handleImageUploaded}
-        handleChange={this.handleChange}
-      />
+      <>
+        <Renderer {...this.props} />
+        {!readOnly ? (
+          <Controls
+            {...this.props}
+            imagePreview={this.state.imagePreview}
+            handleImageLoaded={this.handleImageLoaded}
+            handleImageUploaded={this.handleImageUploaded}
+            handleChange={this.handleChange}
+          />
+        ) : null}
+      </>
     );
   }
 }
