+++ packages/plugins/layout/background/src/index.tsx
@@ -21,11 +21,19 @@
  */
 import { BackgroundSettings } from './types/settings';
 import createPlugin from './createPlugin';
-import BackgroundDefaultControls from './Controls/BackgroundDefaultControls';
+
 import BackgroundHtmlRenderer from './Renderer/BackgroundHtmlRenderer';
 import { MakeOptional } from './types/makeOptional';
 
-export default (settings: MakeOptional<BackgroundSettings, 'Renderer' | 'Controls'>) => {
+import { lazyLoad } from '@react-page/core';
+
+const BackgroundDefaultControls = lazyLoad(() =>
+  import('./Controls/BackgroundDefaultControls')
+);
+
+export default (
+  settings: MakeOptional<BackgroundSettings, 'Renderer' | 'Controls'>
+) => {
   const plugin = createPlugin({
     Controls: BackgroundDefaultControls,
     Renderer: BackgroundHtmlRenderer,
