+++ packages/plugins/layout/parallax-background/src/index.tsx
@@ -22,11 +22,16 @@
 
 import createPlugin from './createPlugin';
 import ParallaxBackgroundHtmlRenderer from './Renderer/ParallaxBackgroundHtmlRenderer';
-import ParallaxBackgroundDefaultControls from './Controls/ParallaxBackgroundDefaultControls';
+
 import { LayoutPluginConfig } from '@react-page/core/lib/service/plugin/classes';
 import { ParallaxBackgroundState } from './types/state';
 import { ParallaxBackgroundSettings } from './types/settings';
 import { MakeOptional } from './types/makeOptional';
+
+import { lazyLoad } from '@react-page/core';
+const ParallaxBackgroundDefaultControls = lazyLoad(() =>
+  import('./Controls/ParallaxBackgroundDefaultControls')
+);
 
 export type ParallaxBackgroundDefaultSettings = MakeOptional<
   ParallaxBackgroundSettings,
