+++ packages/ui/src/ThemeProvider/themeOptions.tsx
@@ -1,0 +1,6 @@
+import { ThemeOptions } from '@material-ui/core/styles/createMuiTheme';
+export const themeOptions: ThemeOptions = {
+  typography: {
+    useNextVariants: true,
+  },
+};
